api/requirements.txt
amqp==2.6.1
argon2-cffi==20.1.0
asgiref==3.3.1
Babel==2.9.1
billiard==3.6.4.0
celery==4.3.0
cffi==1.14.4
colorama==0.4.4
config==0.5.0.post0
Django==3.0
django-cors-headers==3.7.0
django-extensions==3.1.0
django-filter==2.4.0
django-redis==4.10.0
djangorestframework==3.12.2
dnspython==1.16.0
flower==0.9.7
greenlet==1.1.0
humanize==3.9.0
importlib-metadata==4.5.0
kombu==4.6.11
pbr==5.5.1
Pillow==8.1.0
prometheus-client==0.8.0
pycparser==2.20
PyMySQL==1.0.2
PyPDF2==1.26.0
python-decouple==3.4
pytz==2020.5
qrcode==6.1
raven==6.10.0
redis==3.3.8
six==1.15.0
sqlparse==0.4.1
testresources==2.0.1
tornado==5.1.1
typing-extensions==3.10.0.0
vine==1.3.0
zipp==3.4.1
zope.event==4.5.0
zope.interface==5.4.0